backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic[email]
pandas
numpy
scipy
openpyxl
python-jose[cryptography]
bcrypt
python-dotenv
statsmodels
scikit-learn
slowapi
reportlab
psycopg2-binary
pytest
pytest-asyncio
httpx
